move styling helpers to their own file
