+++ package/windows/ReactNativeXaml/Styling.h
@@ -1,0 +1,6 @@
+#pragma once
+#include <JSValue.h>
+
+namespace ColorUtils {
+  xaml::Media::Brush BrushFrom(const winrt::Microsoft::ReactNative::JSValue& v);
+}
